groups-backend/requirements.txt
# Groups Backend - FastAPI Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
supabase>=2.3.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
httpx>=0.24.0
aiofiles>=23.2.1
boto3>=1.34.0
Pillow>=10.2.0
beautifulsoup4>=4.12.2
